refactor: email verification page
